Added Sound Files
Split files into two folders, FX and OST based on if they are a song or a sound effect.

Edited basic scene for sound testing functionality

Sounds changed in fps controller to test footsteps on wood (might need further changes to remove static)
red spheres play sounds, will feature text in future so that you can tell which is which.