reworked the food system so it is filtered, with med system as well
when feeding the animal, one need to consider the food type and the preferred food of the animal
if it matches the food type, it would just increase the hunger, but if it also matches the preferred food, it increases happiness as well.

for the medicine I added the ability to add side effects if needed in the form of +/- hunger and mood, allowing for things like anti-depressants.