Book system fully updated.
Pages have got shading and relevant info now (can be changed in future if needed). Added a help section with general navigation help and some tips on the rules. Also added the tab system that was recommended on Friday. The speed that it turns pages can be edited if we want it to go faster/slower.