Finished inventory UI
Combined my UI with Howard's code to create a working hotbar. Different icons are displayed based on the tag of the picked up item. Really easy to add more variety in item - just create a new tag, add the tag to the string array and then add the icon in the corresponding position in the item array.